Fixed parcel generator algorithm and initiated file download immediately
Previous version of parcel continued splitting left child of each box on
second to last level, thus increasing depth of tree and put large number
of objects on the stack. Fixed this error, which fixed the problem of
generating infinitely small boxes. Additionally it reduces the amount of
time to print the first box.